Refactor: move stats cals out of chart.tsx
Fix: windowDuration stats
