+++ src/components/Chart/data/dataAccumulator.ts
@@ -4,8 +4,6 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-/* eslint no-plusplus: off */
-
 import {
     numberOfDigitalChannels,
     options,
@@ -15,6 +13,41 @@ import bitDataAccumulator, { BitDataAccumulator } from './bitDataAccumulator';
 import { createEmptyArrayWithAmpereState } from './commonBitDataFunctions';
 import { AmpereState, DigitalChannelStates } from './dataTypes';
 import noOpBitDataProcessor from './noOpBitDataProcessor';
+
+export const calcStats = (begin?: null | number, end?: null | number) => {
+    if (begin == null || end == null) {
+        return null;
+    }
+
+    if (end < begin) {
+        [begin, end] = [end, begin];
+    }
+
+    const { data } = options;
+    const indexBegin = Math.ceil(timestampToIndex(begin));
+    const indexEnd = Math.floor(timestampToIndex(end));
+
+    let sum = 0;
+    let len = 0;
+    let max;
+
+    for (let n = indexBegin; n <= indexEnd; n += 1) {
+        const k = (n + data.length) % data.length;
+        const v = data[k];
+        if (!Number.isNaN(v)) {
+            if (max === undefined || v > max) {
+                max = v;
+                sum += v;
+                len += 1;
+            }
+        }
+    }
+    return {
+        average: sum / (len || 1),
+        max: max ?? 0,
+        delta: end - begin,
+    };
+};
 
 export interface DataAccumulator {
     ampereLineData: AmpereState[];
@@ -70,7 +103,7 @@ export default (): DataAccumulator => ({
         for (
             let originalIndex = originalIndexBegin;
             mappedIndex < len + len;
-            ++mappedIndex, originalIndex += step
+            mappedIndex += 1, originalIndex += step
         ) {
             const timestamp =
                 begin + windowDuration * (mappedIndex / (len + len));
@@ -79,7 +112,7 @@ export default (): DataAccumulator => ({
             let min: number | undefined = Number.MAX_VALUE;
             let max: number | undefined = -Number.MAX_VALUE;
 
-            for (let n = k; n < l; ++n) {
+            for (let n = k; n < l; n += 1) {
                 let v = data[n] ?? NaN;
 
                 if (removeZeroValues && v === 0) {
@@ -101,7 +134,7 @@ export default (): DataAccumulator => ({
 
             this.ampereLineData[mappedIndex].x = timestamp;
             this.ampereLineData[mappedIndex].y = min;
-            ++mappedIndex;
+            mappedIndex += 1;
             this.ampereLineData[mappedIndex].x = timestamp;
             this.ampereLineData[mappedIndex].y = max;
